Refactor: network.remote.model 안에 있던 data class 전부 외부 model 폴더로 이전
