+++ app/src/main/java/com/depromeet/housekeeper/model/HouseWorkCreateResponse.kt
@@ -1,4 +1,4 @@
-package com.depromeet.housekeeper.network.remote.model
+package com.depromeet.housekeeper.model
 
 import android.os.Parcelable
 import com.depromeet.housekeeper.model.HouseWork
